+++ inst/exdata/shinyRatingsBinding.js
@@ -7,6 +7,20 @@ $.extend(shinyRatingsInputBinding, {
     return $(scope).find(".shinyRatings");
   },
 
+  initialize: function(el){
+    /* Get default value */
+    var data = JSON.parse(el.getAttribute('data'));
+    var n = data.n;
+    /* Manually set the check property to TRUE for default number of stars */
+    var count = 0
+    $(el).find("input").each(function(inputitem){
+        count = count + 1
+        if(count <= n) {
+        $(this).prop("checked", "true");  
+        }
+    });
+  },
+  
   getValue: function(el) {
     var rate_value = 0;
     $(el).find("input").each(function(inputitem){
@@ -14,16 +28,13 @@ $.extend(shinyRatingsInputBinding, {
           rate_value = parseFloat($(this).val());
       }
     });
-    if (rate_value == 0) {
-      rate_value = 5;
-    }
     return(rate_value)
   },
 
   setValue: function(el, value) {
     el.value = value;
   },
-
+  
   subscribe: function(el, callback) {
     $(el).on('change.shinyRatings', function(event) {
          callback(false);